refactor: handleSubmit 내의 카드 생성 및 로컬 스토리지 반영 함수 분리
